[TIMOB-17876] 64 bit safe code migration
Turn on GCC_WARN_64_TO_32_BIT_CONVERSION flag and fix code for the same

+++ iphone/Classes/KrollContext.h
@@ -62,7 +62,7 @@
 
 #ifdef DEBUG
 // used during debugging only
--(int)queueCount;
+-(NSUInteger)queueCount;
 #endif
 
 -(void)invokeOnThread:(id)callback_ method:(SEL)method_ withObject:(id)obj condition:(NSCondition*)condition_;
