+++ iphone/Classes/TiStreamProxy.h
@@ -17,12 +17,12 @@
 @required
 // DEFINED BEHAVIOR: callback != nil indicates an asynch operation.  length==0 indicates to read all available data into
 // the buffer (and grow it if necessary).  These methods MAY be called by classes other than the TiStreamProxy ducktype (i.e. Ti.Stream module methods)
--(int)readToBuffer:(TiBuffer*)buffer offset:(int)offset length:(int)length callback:(KrollCallback*)callback;
--(int)writeFromBuffer:(TiBuffer*)buffer offset:(int)offset length:(int)length callback:(KrollCallback*)callback;
+-(NSUInteger)readToBuffer:(TiBuffer*)buffer offset:(NSUInteger)offset length:(NSUInteger)length callback:(KrollCallback*)callback;
+-(NSUInteger)writeFromBuffer:(TiBuffer*)buffer offset:(NSUInteger)offset length:(NSUInteger)length callback:(KrollCallback*)callback;
 
 // Used for writeStream/pumping
--(int)writeToStream:(id<TiStreamInternal>)output chunkSize:(int)size callback:(KrollCallback*)callback;
--(void)pumpToCallback:(KrollCallback*)callback chunkSize:(int)size asynch:(BOOL)asynch;
+-(NSUInteger)writeToStream:(id<TiStreamInternal>)output chunkSize:(NSUInteger)size callback:(KrollCallback*)callback;
+-(void)pumpToCallback:(KrollCallback*)callback chunkSize:(NSUInteger)size asynch:(BOOL)asynch;
 
 // Public API : No defined behavior
 -(NSNumber*)isReadable:(id)_void; // PUBLIC API FUNCTION
